Move tests closer to source
