+++ crates/compiler/src/compiler.rs
@@ -123,7 +123,6 @@ struct CompilationIntermediate<'input> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::prelude::*;
 
     #[test]
     fn can_call_compile_empty_without_crash() {
@@ -153,216 +152,5 @@ a {1 + 3} cool expression
             compilation_type: CompilationType::FullCompilation,
             variable_declarations: vec![],
         });
-    }
-
-    #[test]
-    fn populates_string_table() {
-        let file = File {
-            file_name: "test.yarn".to_string(),
-            source: "title: test
----
-foo
-bar
-a {1 + 3} cool expression
-==="
-            .to_string(),
-        };
-        let result = compile(CompilationJob {
-            files: vec![file],
-            library: None,
-            compilation_type: CompilationType::FullCompilation,
-            variable_declarations: vec![],
-        });
-        let string_table = result.string_table;
-        assert_eq!(string_table.len(), 3);
-        assert_eq!(
-            string_table["line:test.yarn-test-0"],
-            StringInfo {
-                text: "foo".to_string(),
-                node_name: "test".to_string(),
-                line_number: 3,
-                file_name: "test.yarn".to_string(),
-                is_implicit_tag: true,
-                metadata: vec![],
-            }
-        );
-        assert_eq!(
-            string_table["line:test.yarn-test-1"],
-            StringInfo {
-                text: "bar".to_string(),
-                node_name: "test".to_string(),
-                line_number: 4,
-                file_name: "test.yarn".to_string(),
-                is_implicit_tag: true,
-                metadata: vec![],
-            }
-        );
-        assert_eq!(
-            string_table["line:test.yarn-test-2"],
-            StringInfo {
-                text: "a {0} cool expression".to_string(),
-                node_name: "test".to_string(),
-                line_number: 5,
-                file_name: "test.yarn".to_string(),
-                is_implicit_tag: true,
-                metadata: vec![],
-            }
-        );
-    }
-
-    #[test]
-    fn catches_expression_errors() {
-        let file = File {
-            file_name: "test.yarn".to_string(),
-            source: "title: test
----
-foo
-bar
-a {very} cool expression
-==="
-            .to_string(),
-        };
-        let result = compile(CompilationJob {
-            files: vec![file],
-            library: None,
-            compilation_type: CompilationType::FullCompilation,
-            variable_declarations: vec![],
-        });
-        assert!(result.program.is_none());
-        let diagnostics = result.diagnostics;
-        assert_eq!(diagnostics.len(), 2);
-
-        let range = Position {
-            line: 5,
-            character: 8,
-        }..=Position {
-            line: 5,
-            character: 9,
-        };
-        let context = "a {very} cool expression\n       ^".to_owned();
-        let first_expected =
-            Diagnostic::from_message("Unexpected \"}\" while reading a function call".to_string())
-                .with_file_name("test.yarn".to_string())
-                .with_range(range.clone())
-                .with_context(context.clone())
-                .with_severity(DiagnosticSeverity::Error);
-
-        let second_expected =
-            Diagnostic::from_message("mismatched input '}' expecting '('".to_string())
-                .with_file_name("test.yarn".to_string())
-                .with_range(range)
-                .with_context(context)
-                .with_severity(DiagnosticSeverity::Error);
-        if diagnostics[0] == first_expected {
-            assert_eq!(diagnostics[1], second_expected);
-        } else {
-            assert_eq!(diagnostics[0], second_expected);
-            assert_eq!(diagnostics[1], first_expected);
-        }
-    }
-
-    #[test]
-    fn recognizes_variable_declarations() {
-        let file = File {
-            file_name: "test.yarn".to_string(),
-            source: "title: test
----
-<<declare $foo to 1>>
-<<declare $bar = \"2\">>
-<<declare $baz to true>>
-<<declare $quux = \"hello there\" as string>>
-==="
-            .to_string(),
-        };
-        let result = compile(CompilationJob {
-            files: vec![file],
-            library: None,
-            compilation_type: CompilationType::FullCompilation,
-            variable_declarations: vec![],
-        });
-
-        println!("{:?}", result.diagnostics);
-        assert!(result.diagnostics.is_empty());
-        let declarations = result.declarations.unwrap();
-        assert_eq!(declarations.len(), 4);
-        assert_eq!(
-            declarations[0],
-            Declaration {
-                name: "$foo".to_string(),
-                default_value: 1.0.into(),
-                description: "".to_string(),
-                source_file_name: DeclarationSource::File("test.yarn".to_string()),
-                source_node_name: Some("test".to_string()),
-                is_implicit: false,
-                r#type: Type::Number(NumberType),
-                range: Position {
-                    line: 3,
-                    character: 11,
-                }..=Position {
-                    line: 3,
-                    character: 15,
-                },
-            }
-        );
-
-        assert_eq!(
-            declarations[1],
-            Declaration {
-                name: "$bar".to_string(),
-                default_value: "2".to_string().into(),
-                description: "".to_string(),
-                source_file_name: DeclarationSource::File("test.yarn".to_string()),
-                source_node_name: Some("test".to_string()),
-                is_implicit: false,
-                r#type: Type::String(StringType),
-                range: Position {
-                    line: 4,
-                    character: 11,
-                }..=Position {
-                    line: 4,
-                    character: 15,
-                },
-            }
-        );
-
-        assert_eq!(
-            declarations[2],
-            Declaration {
-                name: "$baz".to_string(),
-                default_value: true.into(),
-                description: "".to_string(),
-                source_file_name: DeclarationSource::File("test.yarn".to_string()),
-                source_node_name: Some("test".to_string()),
-                is_implicit: false,
-                r#type: Type::Boolean(BooleanType),
-                range: Position {
-                    line: 5,
-                    character: 11,
-                }..=Position {
-                    line: 5,
-                    character: 15,
-                },
-            }
-        );
-
-        assert_eq!(
-            declarations[3],
-            Declaration {
-                name: "$quux".to_string(),
-                default_value: "hello there".to_string().into(),
-                description: "".to_string(),
-                source_file_name: DeclarationSource::File("test.yarn".to_string()),
-                source_node_name: Some("test".to_string()),
-                is_implicit: false,
-                r#type: Type::String(StringType),
-                range: Position {
-                    line: 6,
-                    character: 11,
-                }..=Position {
-                    line: 6,
-                    character: 16,
-                },
-            }
-        );
     }
 }
