+++ crates/compiler/src/visitors/declaration_visitor.rs
@@ -240,3 +240,113 @@ impl<'a, 'input: 'a> YarnSpinnerParserVisitorCompat<'input> for DeclarationVisit
         self.new_declarations.push(declaration);
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn recognizes_variable_declarations() {
+        let file = File {
+            file_name: "test.yarn".to_string(),
+            source: "title: test
+---
+<<declare $foo to 1>>
+<<declare $bar = \"2\">>
+<<declare $baz to true>>
+<<declare $quux = \"hello there\" as string>>
+==="
+            .to_string(),
+        };
+        let result = compile(CompilationJob {
+            files: vec![file],
+            library: None,
+            compilation_type: CompilationType::FullCompilation,
+            variable_declarations: vec![],
+        });
+
+        println!("{:?}", result.diagnostics);
+        assert!(result.diagnostics.is_empty());
+        let declarations = result.declarations.unwrap();
+        assert_eq!(declarations.len(), 4);
+        assert_eq!(
+            declarations[0],
+            Declaration {
+                name: "$foo".to_string(),
+                default_value: 1.0.into(),
+                description: "".to_string(),
+                source_file_name: DeclarationSource::File("test.yarn".to_string()),
+                source_node_name: Some("test".to_string()),
+                is_implicit: false,
+                r#type: Type::Number(NumberType),
+                range: Position {
+                    line: 3,
+                    character: 11,
+                }..=Position {
+                    line: 3,
+                    character: 15,
+                },
+            }
+        );
+
+        assert_eq!(
+            declarations[1],
+            Declaration {
+                name: "$bar".to_string(),
+                default_value: "2".to_string().into(),
+                description: "".to_string(),
+                source_file_name: DeclarationSource::File("test.yarn".to_string()),
+                source_node_name: Some("test".to_string()),
+                is_implicit: false,
+                r#type: Type::String(StringType),
+                range: Position {
+                    line: 4,
+                    character: 11,
+                }..=Position {
+                    line: 4,
+                    character: 15,
+                },
+            }
+        );
+
+        assert_eq!(
+            declarations[2],
+            Declaration {
+                name: "$baz".to_string(),
+                default_value: true.into(),
+                description: "".to_string(),
+                source_file_name: DeclarationSource::File("test.yarn".to_string()),
+                source_node_name: Some("test".to_string()),
+                is_implicit: false,
+                r#type: Type::Boolean(BooleanType),
+                range: Position {
+                    line: 5,
+                    character: 11,
+                }..=Position {
+                    line: 5,
+                    character: 15,
+                },
+            }
+        );
+
+        assert_eq!(
+            declarations[3],
+            Declaration {
+                name: "$quux".to_string(),
+                default_value: "hello there".to_string().into(),
+                description: "".to_string(),
+                source_file_name: DeclarationSource::File("test.yarn".to_string()),
+                source_node_name: Some("test".to_string()),
+                is_implicit: false,
+                r#type: Type::String(StringType),
+                range: Position {
+                    line: 6,
+                    character: 11,
+                }..=Position {
+                    line: 6,
+                    character: 16,
+                },
+            }
+        );
+    }
+}
